requirements.txt
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==5.0.6
django_compressor==4.5
django-browser-reload==1.13.0
openai==1.35.1
pytailwindcss==0.2.0
python-dotenv==1.0.1
Markdown==3.6
requests==2.32.3
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-ratelimit==4.1.0
django-two-factor-auth[phonenumbers]==1.16.0